requirements.txt
playwright
pytest-playwright
allure-pytest
pytest-xdist
